requirements.txt
aiogram==3.2.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
